+++ common.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2019 Expedia Inc.
+ * Copyright (C) 2019 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  */
@@ -10,4 +10,9 @@ locals {
 
 data "aws_vpc" "shuntingyard_vpc" {
   id = "${var.vpc_id}"
+}
+
+data "aws_secretsmanager_secret" "docker_registry" {
+  count = "${var.docker_registry_auth_secret_name == "" ? 0 : 1}"
+  name  = "${var.docker_registry_auth_secret_name}"
 }